ai_tools_bridge/packages/ai_tools_confluence/pyproject.toml
[project]
name = "ai-tools-confluence"
version = "0.1.0"
description = "AI tools for Confluence operations and content management"
readme = "README.md"
authors = [
    { name = "Matthias Rieck", email = "Matthias.Rieck@bmw.de" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.0",
    "ai_tools_base",
    "atlassian-python-api>=4.0.7",
    "html-to-markdown>=1.16.0",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# [tool.hatch.build.hooks.custom]
# path = "hatch_build.py"

# [tool.hatch.version]
# source = "uv-dynamic-versioning"

[dependency-groups]
dev = [
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.12.12",
]

[tool.uv.sources]
ai_tools_base = { path = "../ai_tools_base", editable = true }

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=src --cov-report=term-missing"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.pyright]
include = ["src"]
exclude = ["hatch_build.py"]
typeCheckingMode = "strict"
deprecateTypingAliases = true
reportDeprecated = "error"
reportMissingTypeStubs = false


[[tool.uv.index]]
url = "https://pypi.org/simple"
name = "pypi"
default = true
